requirements.txt
Django>=3.1
requests>=2.24.0
djangorestframework>=3.13.1
psycopg2-binary==2.9.3